+++ src/crypto/test/NIST.CVP.Crypto.AES_FF.Tests/Ff3BlockCipherTests.cs
@@ -191,7 +191,7 @@ namespace NIST.CVP.Crypto.AES_FF.Tests
                 Iv = tweak,
                 Key = key,
                 Payload = NumeralString.ToBitString(payload),
-                Radix = (short)radix
+                Radix = radix
             });
 
             Assert.AreEqual(cipherText.ToString(), NumeralString.ToNumeralString(result.Result).ToString());
@@ -207,7 +207,7 @@ namespace NIST.CVP.Crypto.AES_FF.Tests
                 Iv = tweak,
                 Key = key,
                 Payload = NumeralString.ToBitString(cipherText),
-                Radix = (short)radix
+                Radix = radix
             });
 
             Assert.AreEqual(payload.ToString(), NumeralString.ToNumeralString(result.Result).ToString());
